+++ test/sucrose/sucrose.test.ts
@@ -2,6 +2,7 @@ import { describe, expect, it } from 'bun:test'
 import { Elysia } from '../../src'
 
 import { separateFunction, sucrose } from '../../src/sucrose'
+import { req } from '../utils'
 
 describe('sucrose', () => {
 	it('common 1', () => {
@@ -39,7 +40,32 @@ describe('sucrose', () => {
 			headers: true,
 			body: false,
 			cookie: false,
-			set: false
+			set: false,
+			unknownQueries: false
+		})
+	})
+
+	it('integration 1', async () => {
+		const path = 'a'
+
+		const app = new Elysia()
+			// ✅ easy to perform inference
+			.get('/1', ({ query: { a } }) => a)
+			.get('/2', ({ query }) => query.a)
+			.get('/3', (c) => c.query.a)
+			.get('/4', ({ query }) => query[path])
+			.get('/5', (c) => c.query[path])
+
+		addEventListener('fetch', (request) => {
+			console.log(request)
+		})
+
+		new Array(5).fill(0).map(async (_, i) => {
+			const result = await app
+				.handle(req(`/${i + 1}?a=a&b=b`))
+				.then((x) => x.text())
+
+			expect(result).toBe('a')
 		})
 	})
 })