+++ test/units/map-response.test.ts
@@ -19,6 +19,12 @@ const context = {
 
 class Student {
 	constructor(public name: string) {}
+
+	toString() {
+		return JSON.stringify({
+			name: this.name
+		})
+	}
 }
 
 class CustomResponse extends Response {}